refactor: 변경된 인터페이스 패키지 위치 반영
