+++ server/src/main/java/spring/community/authentication/service/AuthServiceImpl.java
@@ -5,6 +5,9 @@ import org.springframework.stereotype.Service;
 import spring.community.authentication.dto.SignupRequestDto;
 import spring.community.authentication.entity.SignupVerification;
 import spring.community.authentication.repository.SignupVerificationRepository;
+import spring.community.authentication.service.interfaces.AuthDataHelpService;
+import spring.community.authentication.service.interfaces.AuthService;
+import spring.community.authentication.service.interfaces.GenerateAuthKeyService;
 import spring.community.exception.FaultSetBuilderAttributesException;
 import spring.community.mail.MailService;
 import spring.community.user.entity.User;
